Folder structure fix: moved core elements to Core dir
