+++ src/Core/Helpers/validations.js
@@ -30,11 +30,9 @@ export const handleAmountBeforeInput = (e) => {
 
   export const getCurrentDateAndTime = ()=>{
     const today = new Date();
-    const day = String(today.getDate()).padStart(2, '0');
-    const month = String(today.getMonth() + 1).padStart(2, '0');
-    const year = today.getFullYear();
-    const hours = String(today.getHours()).padStart(2, '0');
-    const minutes = String(today.getMinutes()).padStart(2, '0');
-    const formattedDate = `${day}-${month}-${year} ${hours}:${minutes}`;
-    return formattedDate
+    return `${String(today.getDate()).padStart(2,'0')}-
+            ${String(today.getMonth()+1).padStart(2,'0')}-
+            ${String(today.getFullYear()).padStart(2,'0')} 
+            ${String(today.getHours()).padStart(2,'0')}:
+            ${String(today.getMinutes()).padStart(2,'0')}`
   }
