+++ src/Pages/History/HistoryPage.jsx
@@ -5,6 +5,6 @@ import { GlobalContext } from "../../Context/GlobalState";
 export default function HistoryPage(){
     const { editTransaction } = useContext(GlobalContext)
     return(<div className="container">
-        <TransactionList incomeExpenseTransaction = {false} title = "Transaction History" buttonOnClick = {editTransaction}/>
+        <TransactionList incomeExpenseTransaction = {false} title = "Transaction History" buttonOnclick = {editTransaction}/>
     </div>)
 }
